refactor:users and relationships controllers moved into users folder
