+++ spec/requests/api/v1/users/users_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe 'Api::V1::Users', type: :request do
+RSpec.describe 'Api::V1::Users::Users' do
   describe 'GET /index' do
     pending "add some examples (or delete) #{__FILE__}"
   end
